+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Strophe.js Change Log
 
 ## Version 1.2.16 - (Unreleased)
-* Fix websocket close handler exception and reporting
+* #299 'no-auth-mech' error. Server did not offer a supported authentication mechanism
+* #306 Fix websocket close handler exception and reporting
 
 ## Version 1.2.15 - (2018-05-21)
 * #259 XML element should be sent to xmlOutput